Refactor python part for pyodide
