+++ flowfile_wasm/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+# Native pytest suite for the in-browser Pyodide engine (src/pyodide/engine.py).
+# Scoped here so `cd flowfile_wasm && pytest` runs only these and is independent
+# of the monorepo's root pytest configuration.
+testpaths = tests/python
